requirements.txt
numpy
llvmlite
numba
tensorboardX
easydict
pyyaml
scikit-image
tqdm
torchvision
SharedArray
opencv-python
pyquaternion
torchmetrics
torch_scatter
